fix(registry): make connection guard future-base compatible
Relocate guard regressions outside the evolved registry conflict paths and align the overlapping files with the current multi-host base. This keeps the omission/default contract while allowing the reviewed tail-to-root merge order to apply without manual conflict resolution.
