+++ .gitignore
@@ -139,6 +139,3 @@ dist
 build/
 manual-kerberos-test.mjs
 manual-multi-host-test.mjs
-
-# Local scratch TMPDIR (used to dodge full /tmp tmpfs during tests)
-.tmp/
